+++ src/Opprintast.ml
@@ -402,7 +402,7 @@ and core_type1 ctxt f x =
     | Ptyp_variant (l, closed, low) ->
         let first_is_inherit =
           match l with
-          | { Parsetree.prf_desc = Rinherit _ } :: _ -> true
+          | { prf_desc = Rinherit _ } :: _ -> true
           | _ -> false
         in
         let type_variant_helper f x =