+++ src/utils.mli
@@ -1,5 +1,3 @@
-open Ppxlib
-
 val split_at_f : ('a -> bool) -> 'a list -> 'a list * 'a list
 (** [split_at_f f l] is the partition [(l1, l2)] such that [l1] is the longest
     prefix where the predicate [f] holds. The order of the elements is not
@@ -53,27 +51,6 @@ module Fmt : sig
   val rbrace : 'a t
   (** [rbrace ppf ()] is [any "}@\]"] *)
 end
-
-exception TypeCheckingError of string
-exception NotSupported of string
-exception Located of Location.t * exn
-
-val error : loc:Location.t -> exn -> 'a
-(** [error ~loc e] raises [e], wrapped in [Located(loc, e)]. *)
-
-val check : loc:Location.t -> bool -> exn -> unit
-(** [check ~loc b e] checks if [b] is true, otherwise raises [e], wrapped in
-    [Located(loc, e)]. *)
-
-val error_report : loc:Location.t -> string -> 'a
-(** [error_report ~loc e] is [error ~loc (TypeCheckingError s)]. *)
-
-val check_report : loc:Location.t -> bool -> string -> unit
-(** [check_report ~loc b e] is [check ~loc b (TypeCheckingError s)]. **)
-
-val not_supported : loc:Location.t -> string -> 'a
-(** [not_supported ~loc s] raises [NotSupported s], wrapped in
-    [Located(loc, e)]. *)
 
 module Sstr : Set.S with type elt = string
 (** String sets. *)